pd-client: no need to watch leader (#327)
* pd-client: no need to watch leader
* pd-client: delegate a goroutine to reconnect

Since PD server can proxy requests to the leader,
PD client don't need to watch the leader any more.

When we can't connect to the leader, we will connect to any
other servers temporarily. Meanwhile, we delegate a
goroutine to keep trying to connect to the leader. Once the
delegated goroutine connected to the leader, it sends the
connection to the channel, and we can use the leader
connection to replace the old one.

Reference Issues: #293