server: use etcd election for PD leader election (#336)
* server: use etcd election for PD leader election

Here's a brief introduction about how etcd session and election work.

When a session is created, it requests a lease with a ttl, and keep
alive it automatically.

Then we use the session to create an election under a prefix. When we
campaign with a leader value, the election will save the leader value
under the prefix with the session id. For example, if the prefix is
/pd/1/leader, and the session id is 123, then the leader value will be
save at /pd/1/leader/123.  Notice that, multiple elections may save
values under the same prefix concurrently. After the leader value has
been saved, the election will wait until any values created before it
have been deleted. That is, the first created value under the same
prefix is the current leader.

We need to pay attention to one thing. No matter a election becomes
leader or not, it will save its the leader value under the prefix. The
benefit is that if one election exists, the other will become leader
immediately because its leader value has already been saved.

* server/lessor: remove campaign timeout